xauusd-trading-bot-(hmm-architecture)/requirements.txt
# Core
MetaTrader5
pandas
numpy
python-dotenv

# Brain (HMM)
hmmlearn
scikit-learn
scipy

# Dashboard
streamlit
plotly

# Utilities
loguru
schedule
